Added fix for ble not working on rs911x
