+++ src/platform/EFR32/BUILD.gn
@@ -15,6 +15,7 @@
 import("//build_overrides/chip.gni")
 import("//build_overrides/pigweed.gni")
 
+import("${chip_root}/src/platform/device.gni")
 import("${chip_root}/build/chip/buildconfig_header.gni")
 import("${chip_root}/src/crypto/crypto.gni")
 import("${chip_root}/src/platform/device.gni")
@@ -69,12 +70,14 @@ static_library("EFR32") {
     "SystemPlatformConfig.h",
     "wfx_sl_module_init.c",
     "wfx_sl_module_init.h",
-    "rs911x/rsi_ble_config.h",
   ]
 
   if (chip_enable_ble_rs911x) {
     sources += [
       "rs911x/BLEManagerImpl.cpp",
+      "rs911x/wfx_sl_ble_init.h",
+      "rs911x/wfx_sl_ble_init.c",
+      "rs911x/rsi_ble_config.h",
     ]
   } else {
     sources += [ "BLEManagerImpl.cpp" ]
