+++ src/platform/EFR32/rs911x/rsi_ble_config.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
-     * @file  rsi_ble_config.h
+ * @file  rsi_ble_config.h
  * @brief
  *******************************************************************************
  * # License
@@ -19,7 +19,6 @@
 #define RSI_BLE_CONFIG_H
 
 #include "rsi_ble_apis.h"
-#include "rsi_ble_common_config.h"
 #include <rsi_data_types.h>
 /******************************************************
  * *                      Macros
@@ -82,6 +81,8 @@
 
 #define RSI_BLE_MAX_NBR_ATT_SERV 10
 
+#define RSI_BLE_MAX_NBR_MASTERS 1
+#define RSI_BLE_GATT_ASYNC_ENABLE 1
 #define RSI_BLE_GATT_INIT 0
 
 /* Number of BLE GATT RECORD SIZE IN (n*16 BYTES), eg:(0x40*16)=1024 bytes */
